fix(gpu): 14 pre-existing clippy lints under --features ...,vaapi --tests
Surfaced while running clippy --tests with the full feature set during the
hardening pass on 4cbe1f2.  None of these are in production code paths;
they're #[ignore]'d hardware-integration tests + jpeg/test_fixtures
visibility, but they fail clippy -- -D warnings for that feature combo.

- jpeg/test_fixtures: pub(super) → pub on the JPEG byte constants so
  the vaapi sibling can use them (it already imports them; the build
  failed under --features vaapi --tests).  Module is pub(crate), so pub
  on inner items is the clippy-preferred form (avoids
  redundant_pub_crate).
- nvjpeg.rs / nvjpeg2k.rs: three match-on-Result-with-Err→return blocks
  rewritten as let-else (the modern idiom).
- nvjpeg2k.rs: BAD_JPEG backticked in a doc-comment.
- vaapi/mod.rs: use brought to top of test fn (items_after_statements);
  decode_sync inner type added to doc-comment in backticks; two
  unused-must-use Results bound to _; dead `mut` removed from a
  decoder binding whose only call was through an &self method.
- decode_queue.rs: n_threads typed as u32 upfront (was usize cast to
  u32 at the assert site).

+++ crates/gpu/src/decode_queue.rs
@@ -292,7 +292,7 @@ mod tests {
 
         let q = StdArc::new(DecodeQueue::new(AlwaysGray, "test-concurrent"));
         let counter = StdArc::new(AtomicU32::new(0));
-        let n_threads = 8usize;
+        let n_threads = 8u32;
 
         let handles: Vec<_> = (0..n_threads)
             .map(|_| {
@@ -314,7 +314,7 @@ mod tests {
         // Drop the queue explicitly to verify shutdown is clean, matching the
         // production invariant: handles gone before queue drops.
         drop(q);
-        assert_eq!(counter.load(Ordering::Relaxed), n_threads as u32);
+        assert_eq!(counter.load(Ordering::Relaxed), n_threads);
     }
 
     #[test]
